Remove ExtendedTextMetrics flag to launch canvas TextMetrics.
This change removes the flag ExtendedTextMetrics and the code that references
it in order to ship the full set of Text Metrics returned by canvas.measureText.
Currently chrome (as well as safari and edge) is only returning the width of
the measured text. This change exposes the full TextMetrics object defined
by the spec: https://html.spec.whatwg.org/multipage/canvas.html#textmetrics

Intend-to-ship:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/KQW4thKogkk/6fFtjZWvAwAJ

Bug: 277215
Change-Id: I03a6c149f86998cc27abf33816d1a02f7900c492